backend/requirements.txt
neo4j
langchain
langchain-neo4j
langchain-google-genai
langchain-experimental
langchain-core
langchain-text-splitters
langchain-community
langgraph
neomodel
transformers
tqdm
torch
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
websockets
httpx